+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/
 
 - Added Tweaks and Software section to Windows 11 on a Mac page.
 - Moved all images to Imgur now that sphinx-imgur works fine with OpenGraph.
+- Added link to LICENSE file in the left sidebar.
 
 ## 2021-12-13